build: clear the warnings outside api/brep
A clean build reported 72 at MSVC's default /W1. Three files held 51 of them.

api/font/TextureFont.cxx wrote the 4x4 white block a line glyph is drawn from
as -1 in an unsigned char array, 48 narrowing conversions in one initialiser.
0xff is the same bytes and says what it means.

moya's RenderContext.cxx was a real defect rather than noise: glm::scale and
glm::translate return the transformed matrix instead of modifying in place, so
discarding both returns left the raster coordinate system as the identity. The
call order already expressed the intended screen-to-raster mapping, so
assigning the results is the whole fix - which is what [[nodiscard]] was
telling us. RenderMan.cxx narrows va_arg's promoted double to the float it
assigns explicitly.

The 21 that remain are all in api/brep/BRep.cxx and are the size_t/unsigned
int item on docs/TODO.md rather than this one: BRep indexes with unsigned int,
HalfEdge stores uint64_t, and addVertex/addEdge/addFace return size_t, so the
conversions are the design rather than an oversight. Left alone here.

+++ moya/libmoya/RenderMan.cxx
@@ -311,7 +311,7 @@ RtVoid RiProjection(RtToken name, ...) {
         if (param != RI_NULL) {
             token = va_arg(ap, RtToken);
             if (token == RI_FOV) {
-                fov = va_arg(ap, double);
+                fov = static_cast<float>(va_arg(ap, double));
             }
         }
         _renderer.activeRenderContext().projection(name, fov);
